requirements.txt
flask
google-cloud-speech
